I dont know what is this thats a lot of things happened
But grappling works, and the rest is like trying to fix movement or something.
Trying to implement Checkpoint which doesn't work and implement Level Reset but it is making player HP not working properly. Respawning work but it also make player HP not working properly.
Implement dont destroy on load for all manager stuff I guess (not sure this work the intended way or not)